% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMeteoImputedMissing.R
\name{AddMeteoImputedMissing}
\alias{AddMeteoImputedMissing}
\title{AddMeteoImputedMissing}
\usage{
AddMeteoImputedMissing(weather_data_dep, weather_data_reg)
}
\arguments{
\item{weather_data_dep}{The departement weather dataset. Usually from PredictCovid::weather_data_dep.}

\item{weather_data_reg}{The region weather dataset. Usually from PredictCovid::weather_data_reg.}
}
\value{
The imputed dataframe with meteo data.
}
\description{
Create a dataset with imputed meteo dataset
}
\examples{
\dontrun{

df_hospitPCRDMMeteo <- AddMeteoImputedMissing(weather_data_dep = PredictCovid::weather_data_dep,
                                              weather_data_reg = PredictCovid::weather_data_reg)
}
}
