% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrangeModelbyCompromise.R
\name{ArrangeModelbyCompromise}
\alias{ArrangeModelbyCompromise}
\title{ArrangeModelbyCompromise}
\usage{
ArrangeModelbyCompromise(dfPerf, forecast = c(4, 7, 14))
}
\arguments{
\item{dfPerf}{The performance dataframe from Dfperformances()}

\item{forecast}{The forecast times}
}
\value{
A dataframe with model performance arrange by sum of MRE
}
\description{
Find best model (sum of MRE) on all forecast time
}
