% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddf_modelPCR.R
\name{builddf_modelPCR}
\alias{builddf_modelPCR}
\title{builddf_modelPCR}
\usage{
builddf_modelPCR(df, biaswindow = 7, forecastwindow = 7)
}
\arguments{
\item{df}{Base dataframe usually df_pcrAfterMay from datagouv_importData.}

\item{biaswindow}{Cumulate the information on multiple days to take into account the measurement bias. Default is 7 days.}

\item{forecastwindow}{The forecast distance for prediction. Default is 7 days.}
}
\value{
A dataframe with forecast the variable to forecast and explanatory variables.
}
\description{
Build dataset to forecast number of COVID positive cases
}
