% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDfPredictions.R
\name{BuildDfPredictions}
\alias{BuildDfPredictions}
\title{BuildDfPredictions}
\usage{
BuildDfPredictions(dfDataPred, dfDashboard)
}
\arguments{
\item{dfDataPred}{The raw prediction dataframe}

\item{dfDashboard}{The raw data}
}
\value{
A cleaned prediction dataframe
}
\description{
Build the prediction dataframe from predictions and raw values (last one is used for constant predictions)
}
