% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDfPredictionVariation.R
\name{BuildDfPredictionVariation}
\alias{BuildDfPredictionVariation}
\title{BuildDfPredictionVariation}
\usage{
BuildDfPredictionVariation(dfDataPred, dfDashboard)
}
\arguments{
\item{dfDataPred}{The raw prediction dataframe}

\item{dfDashboard}{The raw data}
}
\value{
A cleaned prediction dataframe
}
\description{
Build the prediction variation dataframe from predictions and raw values (last one is used for constant predictions)
}
