% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHULmModel.R
\name{CHULmModel}
\alias{CHULmModel}
\title{CHULmModel}
\usage{
CHULmModel(param_FORECAST, param_DATE, Features, dfCHU, variation = F, mfp = F)
}
\arguments{
\item{param_FORECAST}{Forecast time}

\item{param_DATE}{Date from which prediction is made}

\item{Features}{Selected features}

\item{dfCHU}{dataframe with features}

\item{variation}{should the model work on variation rather than raw number}

\item{mfp}{should fractional polynomial be used on numeric values}
}
\value{
A list of two dataframes: prediction and coefficients
}
\description{
Train a model on CHU data with SIDEP predictions
}
