% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanCOVID_FEATURES_SAVE_BIS.R
\name{CleanCOVID_FEATURES_SAVE_BIS}
\alias{CleanCOVID_FEATURES_SAVE_BIS}
\title{CleanCOVID_FEATURES_SAVE_BIS}
\usage{
CleanCOVID_FEATURES_SAVE_BIS(RqCOVID_FEATURES_SAVE_BIS)
}
\arguments{
\item{RqCOVID_FEATURES_SAVE_BIS}{The table request}
}
\value{
A cleaned table
}
\description{
Clean the table from COVID_FEATURES_SAVE_BIS
}
