% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanCOVID_INDIC.R
\name{CleanCOVID_INDIC}
\alias{CleanCOVID_INDIC}
\title{CleanCOVID_INDIC}
\usage{
CleanCOVID_INDIC(RqCOVID_INDIC)
}
\arguments{
\item{RqCOVID_INDIC}{The table from COVID_INDIC table.}
}
\value{
A cleaned dataframe
}
\description{
Clean data from RqCovid_Indic table
}
