% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanEDSOpenData.R
\name{CleanEDSOpenData}
\alias{CleanEDSOpenData}
\title{CleanEDSOpenData}
\usage{
CleanEDSOpenData(dfEDS, dfOpen, delayEDSOpen = 4)
}
\arguments{
\item{dfEDS}{The data from EDS}

\item{dfOpen}{The data from Open data}

\item{delayEDSOpen}{The delay between both sources}
}
\value{
A list of length two with both cleaned dataframe
}
\description{
Clean open data dataframe and EDS dataframe.
}
