% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanLabelBestModel.R
\name{CleanLabelBestModel}
\alias{CleanLabelBestModel}
\title{CleanLabelBestModel}
\usage{
CleanLabelBestModel(model)
}
\arguments{
\item{model}{A model ID in dataframe format}
}
\value{
A character string of model
}
\description{
Take a model id as dataframe and change it to a nice character vector
}
