% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeModelEffects.R
\name{ComputeModelEffects}
\alias{ComputeModelEffects}
\title{ComputeModelEffects}
\usage{
ComputeModelEffects(
  model,
  dfPerf,
  dflsDataGroupImp,
  dflsDataImp,
  dfDashboardRaw,
  DateofInterest
)
}
\arguments{
\item{model}{The model of interest}

\item{dfPerf}{The performance dataframe}

\item{dflsDataGroupImp}{The ID_param model corresponding}

\item{dflsDataImp}{The importance merged to dfPerf using dflsDataGroupImp}

\item{dfDashboardRaw}{The raw features}

\item{DateofInterest}{The date of interest}
}
\value{
A dataframe with model effect.
}
\description{
Compute the model parameter effect (value * coefficient)
}
