% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeOutcome.R
\name{ComputeOutcome}
\alias{ComputeOutcome}
\title{ComputeOutcome}
\usage{
ComputeOutcome(df, OUTCOME, FORECAST, outcomeCol = "CHU_HOSP")
}
\arguments{
\item{df}{Dataframe}

\item{OUTCOME}{The OUTCOME must be one of HOSP or IN_HOSP or HOSP_DERIV or IN_HOSP_in_COUNT_rolMean or RAW_HOSP}

\item{FORECAST}{The forecast time.}

\item{outcomeCol}{The outcome column. Default is 'CHU_HOSP'. Only works for OUTCOME in 'HOSP', 'HOSP_DERIV' or 'RAW_HOSP'.}
}
\value{
The dataframe with the outcome
}
\description{
Compute the outcome on the given dataframe
}
