% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeOutcomeCustom.R
\name{ComputeOutcomeCustom}
\alias{ComputeOutcomeCustom}
\title{ComputeOutcomeCustom}
\usage{
ComputeOutcomeCustom(df, OUTCOME, FORECAST)
}
\arguments{
\item{df}{Dataframe}

\item{OUTCOME}{The OUTCOME must be one of HOSP or IN_HOSP}

\item{FORECAST}{The forecast time.}
}
\value{
The dataframe with the outcome
}
\description{
Compute the custom outcome on the given dataframe
}
