% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeOutcomeRef.R
\name{ComputeOutcomeRef}
\alias{ComputeOutcomeRef}
\title{ComputeOutcomeRef}
\usage{
ComputeOutcomeRef(df, OUTCOME, FORECAST, outcomeCol = "CHU_HOSP")
}
\arguments{
\item{df}{Dataframe}

\item{OUTCOME}{The OUTCOME must be one of HOSP or IN_HOSP}

\item{FORECAST}{The forecast time.}

\item{outcomeCol}{The outcome column. Default is 'CHU_HOSP'. Only works for OUTCOME in 'HOSP', 'HOSP_DERIV' or 'RAW_HOSP'.}
}
\value{
The dataframe with the outcome
}
\description{
Compute the outcome of reference on the given dataframe
}
