% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputerolMeanMaxMinDeriv.R
\name{ComputerolMeanMaxMinDeriv}
\alias{ComputerolMeanMaxMinDeriv}
\title{ComputerolMeanMaxMinDeriv}
\usage{
ComputerolMeanMaxMinDeriv(
  df,
  date_column = "START_DATE",
  excludeTansf = "START_DATE"
)
}
\arguments{
\item{df}{The dataset}

\item{date_column}{The date column. Default is 'START_DATE'.}

\item{excludeTansf}{The double column not to take into account.}
}
\value{
The dataframe with the transformed features
}
\description{
Function to compute rolMean, rolMin, rolMax and rolDeriv on double columns of a dataset
}
