% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsecutiveNa.R
\name{ConsecutiveNa}
\alias{ConsecutiveNa}
\title{ConsecutiveNa}
\usage{
ConsecutiveNa(df, vars_to_check)
}
\arguments{
\item{df}{The dataframe of interest.}

\item{vars_to_check}{The features to check.}
}
\value{
A dataframe with the max consecutive NA by dep and by feature.
}
\description{
Check the max number of consecutive NA by dep among the vars_to_check.
}
