% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovidDataFromI2b2.R
\name{CovidDataFromI2b2}
\alias{CovidDataFromI2b2}
\title{CovidDataFromI2b2}
\usage{
CovidDataFromI2b2(RqCOVID_INDIC, RqCOVID_FEATURES_SAVE_BIS)
}
\arguments{
\item{RqCOVID_INDIC}{Data from COVID_INDIC table}

\item{RqCOVID_FEATURES_SAVE_BIS}{Data from COVID_FEATURES_SAVE_BIS table}
}
\value{
A table with all the data
}
\description{
Aggregate data from COVID_INDIC and COVID_FEATURES_SAVE_BIS i2b2 tables
}
