% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvFrechetRf_slarTask.R
\name{cvFrechetRf_slarTask}
\alias{cvFrechetRf_slarTask}
\title{cvFrechetRf_slarTask}
\usage{
cvFrechetRf_slarTask(
  df,
  char_colDate = "DATE",
  slar_taskid,
  dbl_curvePastWindow = 14,
  ncores = 2,
  ntree = 100,
  subset_training = F,
  param_YTYPE,
  param_FORECAST,
  param_DATE,
  param_FE,
  delaySIDEP = 0
)
}
\arguments{
\item{df}{Dataframe}

\item{char_colDate}{The date column name (default is "DATE")}

\item{slar_taskid}{Slurm task id}

\item{dbl_curvePastWindow}{Curve past window (default is 14)}

\item{ncores}{cores number (default is 2)}

\item{ntree}{Number of tree for randomforest (default is 100)}

\item{subset_training}{learn on 1 out of ten days (default is F). If TRUE, learning will be faster but possibly less accurate}

\item{param_YTYPE}{a character vector of length 1 indicating the type of Y variable (scalar or curve)}

\item{param_FORECAST}{a dbl vector of length 1 indicating the forecast time}

\item{param_DATE}{a date vetor of length 1 indicating the date of which test prediction is made}

\item{param_FE}{a character vector describing if feature ingeneered variables should be used or not}

\item{delaySIDEP}{a numeric scalar, SIDEP delay with CHU, 0 (default) if no delay. Usually there are 4 days delay.}
}
\value{
A list of two dataframe: cross validated tree predictions and variables importance
}
\description{
Compute cross validated performance of Frechet randomforest model adapted to slurm curta
}
