% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvlm_slarTask.R
\name{cvlm_slarTask}
\alias{cvlm_slarTask}
\title{cvlm_slarTask}
\usage{
cvlm_slarTask(dbl_depOfInterest = 33, df, char_colDate = "DATE", slar_taskid)
}
\arguments{
\item{dbl_depOfInterest}{Depatement of interest (default is 33)}

\item{df}{Dataframe}

\item{char_colDate}{The date column name (default is "DATE")}

\item{slar_taskid}{Slurm task id}
}
\value{
A list of two dataframe: cross validated tree predictions and variables importance
}
\description{
Compute cross validated performance of linear regression model
}
