% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvRf_slarTask.R
\name{cvRf_slarTask}
\alias{cvRf_slarTask}
\title{cvRf_slarTask}
\usage{
cvRf_slarTask(
  df,
  char_colDate = "DATE",
  slar_taskid,
  ntree = 1,
  param_Y = param_Y,
  param_FORECAST = param_FORECAST,
  param_DATE = param_DATE,
  param_X = param_X,
  delaySIDEP = NULL,
  param_VARIATION = F,
  param_METEO = F,
  param_NPI = -1,
  param_SMOOTH = 0,
  outcome = "hosp",
  RfFull = T,
  EnetFull = T,
  EnetSubset = T,
  RfSubset = T,
  EnetSubset33SIDEP = T,
  EnetSubset33CHU = T,
  variableSubset = c("hosp", "hosp_fct_firstDeriv7", "dep",
    "hosp_fct_firstDeriv7_fct_firstDeriv7", "Population", "P_tous_ages",
    "P_tous_ages_fct_firstDeriv7", "P_tous_ages_fct_firstDeriv7_fct_firstDeriv7",
    "FracP_tous_ages_fct_firstDeriv7", "FracP_tous_ages_fct_firstDeriv7_fct_firstDeriv7",
    "FracP_tous_ages", "IPTCC.mean_avg7"),
  varNPI = c("conf1", "conf2", "ecole", "bar", "mas_clo", "mas_pub_clo", "X6pm",
    "X8pm", "X9pm")
)
}
\arguments{
\item{df}{Dataframe}

\item{char_colDate}{The date column name (default is "DATE")}

\item{slar_taskid}{Slurm task id}

\item{ntree}{Number of tree for randomforest (default is 1)}

\item{param_Y}{a character vector of length 1 indicating the type of Y variable}

\item{param_FORECAST}{a dbl vector of length 1 indicating the forecast time}

\item{param_DATE}{a date vetor of length 1 indicating the date of which test prediction is made}

\item{param_X}{a character vector of length 1 indicating the X variables included}

\item{delaySIDEP}{a numeric scalar, SIDEP delay with CHU, NULL (default) if no delay. Usually there are 4 days delay.}

\item{param_VARIATION}{a boolean, work on variation of number (ratio) instead of raw values.}

\item{param_METEO}{boolean (default is FALSE). Should meteorological data be used.}

\item{param_NPI}{numeric (default is -1). Should NPI be used for prediction and with which delay.}

\item{param_SMOOTH}{numeric (default is 0). Should derivatives and meanCases be calculated on a smoothed feature value.}

\item{outcome}{The outcome of interest, default is "hosp".}

\item{RfFull}{boolean. Train randomforest on all variables.}

\item{EnetFull}{boolean. Train elastic-net on all variables.}

\item{EnetSubset}{boolean. Train elastic-net on subset variables.}

\item{RfSubset}{boolean. Train randomforest on subset variables.}

\item{EnetSubset33SIDEP}{boolean. Train elastic-net on subset variables with only 33 data.}

\item{EnetSubset33CHU}{boolean. Train elastic-net on subset variables with only CHU data.}

\item{variableSubset}{A character vector. Default is c("hosp", "hosp_fct_firstDeriv7", "dep", "Population", "P_tous_ages", "P_tous_ages_fct_firstDeriv7", "FracP_tous_ages_fct_firstDeriv7", "FracP_tous_ages", "IPTCC.mean_avg7")}

\item{varNPI}{A character vector. Default is c("conf1", "conf2", "ecole", "bar", "mas_clo", "mas_pub_clo", "X6pm", "X8pm", "X9pm")}
}
\value{
A list of two dataframe: cross validated tree predictions and variables importance
}
\description{
Compute cross validated performance of randomforest model adapted to slurm curta
}
