% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvRfpred.R
\name{cvRfpred}
\alias{cvRfpred}
\title{cvRfpred}
\usage{
cvRfpred(
  dbl_depOfInterest = 33,
  df,
  char_colDate = "DATE",
  date_min = as_date("2020-08-01", format = "\%Y-\%m-\%d"),
  ntree = 1
)
}
\arguments{
\item{dbl_depOfInterest}{Depatement of interest (default is 33)}

\item{df}{Dataframe}

\item{char_colDate}{The date column name (default is "DATE")}

\item{date_min}{The minimum date for learning before cross-validation (default 2020-08-01)}

\item{ntree}{Number of tree for randomforest (default is 1)}
}
\value{
A dataframe with cross-validated predictions
}
\description{
Compute cross validated performance of randomforest model
}
