% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagouv_importData.R
\name{datagouv_importData}
\alias{datagouv_importData}
\title{datagouv_importData}
\usage{
datagouv_importData(
  url_classeAge09 = "https://bit.ly/30YjDOW",
  url_classeAgeABC = "https://bit.ly/2NwNGKq",
  url_labelRegion = "https://bit.ly/3cMbrH7",
  url_pcrIncDepAfterMay = "https://bit.ly/3eXAt90",
  url_pcrIncDepBeforeMay = "https://bit.ly/3ltLnV8",
  url_hospitPrevDep = "https://bit.ly/3eT8pU4",
  url_opencovid19frChiffrescles = "https://bit.ly/30VbjPU",
  clean = T,
  vec_reliableSources = c("agences-regionales-sante", "sante-publique-france",
    "ministere-sante", "sante-publique-france-data", "opencovid19-fr")
)
}
\arguments{
\item{url_classeAge09}{Age class labels for 0-9. Default is https://www.data.gouv.fr/en/datasets/r/2acb510e-1619-4d5e-8e33-6abd5b4be455}

\item{url_classeAgeABC}{Age class labels for ABC. Default is https://www.data.gouv.fr/en/datasets/r/db378f2a-83a1-40fd-a16c-06c4c8c3535d}

\item{url_labelRegion}{Region labels. Default is https://www.data.gouv.fr/en/datasets/r/0f5643d8-026d-4532-83c6-96e68d429f56}

\item{url_pcrIncDepAfterMay}{PCR dataset after May. Default is https://www.data.gouv.fr/en/datasets/r/406c6a23-e283-4300-9484-54e78c8ae675}

\item{url_pcrIncDepBeforeMay}{PCR dataset before May. Default is https://www.data.gouv.fr/en/datasets/r/b4ea7b4b-b7d1-4885-a099-71852291ff20}

\item{url_hospitPrevDep}{Hospitalisations dataset. Default is https://www.data.gouv.fr/en/datasets/r/08c18e08-6780-452d-9b8c-ae244ad529b3}

\item{url_opencovid19frChiffrescles}{Opencovid dataset. Default is https://raw.githubusercontent.com/opencovid19-fr/data/master/dist/chiffres-cles.csv}

\item{clean}{If clean is equal to true then region column (reg) is added to each datasets. Dep 975, 977, 978 are excluded from the PCR after May dataset. Only reliable ressources are taken from opencovid. Default is TRUE.}

\item{vec_reliableSources}{Vector of reliable ressources from opencovid. Only used if clean is T. Default is c("agences-regionales-sante", "sante-publique-france", "ministere-sante", "sante-publique-france-data", "opencovid19-fr")}
}
\value{
A list of four datasets: PCR after and before may, hospitalisations and opencovid.
}
\description{
Function to import covid datagouv of Santé Publique France (PCR before and after May, Hospitalisations) and opencovid.fr
}
