% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagouv_importHospitPCR.R
\name{datagouv_importHospitPCR}
\alias{datagouv_importHospitPCR}
\title{datagouv_importHospitPCR}
\usage{
datagouv_importHospitPCR(
 
    url_pcrIncDepAfterMay = "https://www.data.gouv.fr/fr/datasets/r/406c6a23-e283-4300-9484-54e78c8ae675",
 
    url_hospitPrevDep = "https://www.data.gouv.fr/fr/datasets/r/63352e38-d353-4b54-bfd1-f1b3ee1cabd7",
 
    url_hospitIncDep = "https://www.data.gouv.fr/fr/datasets/r/6fadff46-9efd-4c53-942a-54aca783c30c",
  df_chu = NULL,
  hospitIncident = FALSE
)
}
\arguments{
\item{url_pcrIncDepAfterMay}{PCR dataset after May. Default is https://www.data.gouv.fr/en/datasets/r/406c6a23-e283-4300-9484-54e78c8ae675}

\item{url_hospitPrevDep}{Hospitalisations dataset. Default is https://www.data.gouv.fr/en/datasets/r/08c18e08-6780-452d-9b8c-ae244ad529b3}

\item{url_hospitIncDep}{Hospitalisations incidence dataset. Default is https://www.data.gouv.fr/fr/datasets/r/6fadff46-9efd-4c53-942a-54aca783c30c}

\item{df_chu}{A dataframe of one or more Hospitals. Each hospital is considered as a new dataframe Must pass quality check of dimensions and column names.}

\item{hospitIncident}{If TRUE, the number of incident hospitalisation over the last 7 days is retrieved instead of the prevalent hospitalisations.}
}
\value{
A dataframe with number of tested and covid positive people by age and number of people in hospitalisation and reanimation
}
\description{
Function to import PCR and hospitalisation covid datagouv of Sante Publique France
}
\examples{
\dontrun{
df_hospitPCR <- datagouv_importHospitPCR()
}
}
