% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagouvJeuxDeDonnees.R
\name{datagouvJeuxDeDonnees}
\alias{datagouvJeuxDeDonnees}
\title{datagouvJeuxDeDonnees}
\usage{
datagouvJeuxDeDonnees(url_listesJeuxDeDonnees = "https://bit.ly/3sfCOj2")
}
\arguments{
\item{url_listesJeuxDeDonnees}{default is https://www.data.gouv.fr/fr/organizations/sante-publique-france/datasets.csv}
}
\value{
A dataframe with informations of different tables available. More informations at https://www.data.gouv.fr/fr/organizations/sante-publique-france/
}
\description{
Import dataset of covid data available on datagouv.fr
}
\examples{
\dontrun{
df_jeuxDeDonnees <- datagouvJeuxDeDonnees()
}
}
