% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateAfterReachValue.R
\name{DateAfterReachValue}
\alias{DateAfterReachValue}
\title{DateAfterReachValue}
\usage{
DateAfterReachValue(value, df, feature, date = "START_DATE")
}
\arguments{
\item{value}{A numeric value}

\item{df}{a dataframe}

\item{feature}{the feature column name}

\item{date}{the date column name}
}
\value{
A date
}
\description{
Obtain date after a feature exceed a certain threshold value.
}
