% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_correspondanceDptRegion.R
\docType{data}
\name{df_dptRegion}
\alias{df_dptRegion}
\title{Correspondance département région et population départementale}
\format{
An object of class \code{data.frame} with 101 rows and 3 columns.
}
\source{
\href{https://www.insee.fr/fr/statistiques/fichier/4316069/cog_ensemble_2020_csv.zip}{INSEE - régions}

\href{https://www.insee.fr/fr/statistiques/1893198}{INSEE - population}
}
\usage{
data(df_dptRegion)
}
\description{
Correspondance département région et population départementale
}
\keyword{datasets}
