% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfClassesAgeVariants.R
\docType{data}
\name{dfClassesAgeVariants}
\alias{dfClassesAgeVariants}
\title{COVID19 variants age class}
\format{
A data frame with 11 rows and 2 columns
}
\usage{
data(dfClassesAgeVariants)
}
\description{
The dataframe contains the following columns:
\itemize{
  \item Classe_age The corresponding age class
  \item indicateur_classe The class indicator
}
}
\keyword{datasets}
