% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfNPI.R
\docType{data}
\name{dfNPI}
\alias{dfNPI}
\title{COVID19 Non pharmaceutical interventions}
\format{
A data frame with 4404 rows and 13 variables
}
\usage{
data(dfNPI)
}
\description{
The dataframe contains the following columns:
\itemize{
  \item admin_name The SEIRcovid region id
  \item date_day The date
  \item conf1 The first lockdown
  \item conf2 The second lockdown
  \item ecole Closed schools
  \item bar Closed pub
  \item mas_clo Mask
  \item mas_pub_clo Mask public
  \item X6pm Curfew at 6 pm
  \item X8pm Curfew at 8 pm
  \item X9pm Curfew at 9 pm
  \item label_insee The insee label
  \item code_insee The insee code
}
}
\examples{
\dontrun{
library(dplyr)
dfNPIraw <- read.table("extdata/mesures_covid/DataSetClimateAndNPI modif.txt")

dfNPIInsee <- data.frame(admin_name = unique(dfNPIraw$admin_name),
                        label_insee = c("Île-de-France",
                                        "Centre-Val de Loire",
                                        "Bourgogne-Franche-Comté",
                                        "Normandie", "Hauts-de-France",
                                        "Grand Est",
                                        "Pays de la Loire", "Bretagne",
                                        "Nouvelle-Aquitaine",
                                        "Occitanie", "Auvergne-Rhône-Alpes",
                                        "Provence-Alpes-Côte d'Azur")) \%>\%
 left_join(PredictCovid::weather_data_reg \%>\%
             select(label_insee, code_insee) \%>\%
             distinct())

dfNPI <- dfNPIraw \%>\%
 select(admin_name, date_day, conf1,
        conf2, ecole, bar, mas_clo,
        mas_pub_clo, X6pm, X8pm,
        X9pm) \%>\%
 left_join(dfNPIInsee) \%>\%
 mutate(date_day = as.Date(date_day))

save(dfNPI, file = "data/dfNPI.rdata")
}
}
\keyword{datasets}
