% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dfperformances.R
\name{Dfperformances}
\alias{Dfperformances}
\title{Dfperformances}
\usage{
Dfperformances(df)
}
\arguments{
\item{df}{The dataframe to compute performance}
}
\value{
A dataframe with MAE and MRE
}
\description{
Compute the performance dataframe with MAE and MRE
}
