% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfperformancesSensitivityCluster.R
\name{DfperformancesSensitivityCluster}
\alias{DfperformancesSensitivityCluster}
\title{DfperformancesSensitivityCluster}
\usage{
DfperformancesSensitivityCluster(df, dfSkipCluster, dfSidepLmPred)
}
\arguments{
\item{df}{The dataframe to compute performance}

\item{dfSkipCluster}{The dataframe without cluster}

\item{dfSidepLmPred}{The SIDEP data}
}
\value{
A dataframe with MAE and MRE
}
\description{
Compute the performance dataframe with MAE and MRE. Sensitivity analysis on SIDEP
}
