% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfperformancesSensitivitySIDEP.R
\name{DfperformancesSensitivitySIDEP}
\alias{DfperformancesSensitivitySIDEP}
\title{DfperformancesSensitivitySIDEP}
\usage{
DfperformancesSensitivitySIDEP(df)
}
\arguments{
\item{df}{The dataframe to compute performance}
}
\value{
A dataframe with MAE and MRE
}
\description{
Compute the performance dataframe with MAE and MRE. Sensitivity analysis on SIDEP
}
