% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DfSidepPredictions.R
\name{DfSidepPredictions}
\alias{DfSidepPredictions}
\title{DfSidepPredictions}
\usage{
DfSidepPredictions(dfPred)
}
\arguments{
\item{dfPred}{The prediction data frame.}
}
\value{
A clean sidep CHU prediction dataframe
}
\description{
Generate a clean sidep CHU prediction dataframe
}
