% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_createVar.R
\name{dm_createVar}
\alias{dm_createVar}
\title{dm_createVar}
\usage{
dm_createVar(df, ls_fctPastCum)
}
\arguments{
\item{df}{The dataframe of interest (obtain with datagouv_importHospitPCR)}

\item{ls_fctPastCum}{The functions for past cumulation (obtain with ls_fctmeanCasesDeriv)}
}
\value{
A dataframe with the newly created variables
}
\description{
Datamanage variables for prediction.
}
