% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddf_modelPCR.R
\name{fct_firstDeriv}
\alias{fct_firstDeriv}
\title{fct_firstDeriv}
\usage{
fct_firstDeriv(x, lagtime)
}
\arguments{
\item{x}{vector to take derivative of}

\item{lagtime}{deepness of lag}
}
\value{
a vector of derivatives
}
\description{
internal function for builddf_modelPCR
}
