% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddf_modelPCR.R
\name{fct_meanCases}
\alias{fct_meanCases}
\title{fct_meanCases}
\usage{
fct_meanCases(x, lagtime)
}
\arguments{
\item{x}{vector to cumsum lag of}

\item{lagtime}{deepness of lag}
}
\value{
a vector of lag cumsum
}
\description{
internal function for builddf_modelPCR
}
