% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testqualitydf.R
\name{fct_testqualitydf}
\alias{fct_testqualitydf}
\title{fct_testqualitydf}
\usage{
fct_testqualitydf(df_to_test, hospitIncident = FALSE)
}
\arguments{
\item{df_to_test}{A dataframe with th following features:
\itemize{
  \item 'dep' [character] - the departement (e.g 33)
  \item 'DATE' [POSIXct] - the date in POSIXct format
  \item 'P_0_9_ans', 'P_10_19_ans', ..., 'P_80_89_ans' and 'P_90_ans_et_plus' [integer] - number of PCR positive patients by age class.
  \item 'P_tous_ages' [integer] - total number of PCR positive patients.
  \item 'TESTED_0_9_ans', 'TESTED_10_19_ans', ..., 'TESTED_80_89_ans' and 'TESTED_90_ans_et_plus' [integer] - number of tested patients by age class.
  \item 'TESTED_tous_ages' [integer] - total number of tested patients.
  \item 'hosp' [integer] - total number of hospitalised patients (current hospitalisations).
  \item 'FROM' [character] - A character string to denote from where the data come from (e.g SIDEP, CHU, ...)
}}

\item{hospitIncident}{A boolean, TRUE if the dataframe is about incidence cases (default is FALSE).}
}
\value{
An error message if something is wrong.
}
\description{
Function to test quality of dataframe for covid-19 prediction
}
