% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctCleanFeaturesName.R
\name{FctCleanFeaturesName}
\alias{FctCleanFeaturesName}
\title{FctCleanFeaturesName}
\usage{
FctCleanFeaturesName(vecFeatures)
}
\arguments{
\item{vecFeatures}{A character vector of features names}
}
\value{
A character vector with the cleaned features names.
}
\description{
Clean features names (more explicit names, convert French to English)
}
