% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctComputeY.R
\name{FctComputeY}
\alias{FctComputeY}
\title{FctComputeY}
\usage{
FctComputeY(df, param_Y, param_FORECAST, outcome = "hosp")
}
\arguments{
\item{df}{A dataframe with 'dep' column the department ID and 'DATE' column, the DATE column}

\item{param_Y}{The Y of interest}

\item{param_FORECAST}{The forecast distance}

\item{outcome}{The outcome of interest, default is "hosp".}
}
\value{
Add the outcome, the outcomeModel (different if HOSPHAB is used) and the outcomeDate
}
\description{
Function to compute the forecast Y with a param_FORECAST distance
}
