% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctPredictionGlmnet.R
\name{FctPredictionGlmnet}
\alias{FctPredictionGlmnet}
\title{FctPredictionGlmnet}
\usage{
FctPredictionGlmnet(
  modelEnet,
  xTest,
  fullTest,
  param_Y,
  param_VARIATION,
  s = "lambda.1se"
)
}
\arguments{
\item{modelEnet}{An elastic net model from cv.glmnet}

\item{xTest}{A test set}

\item{fullTest}{The full test set with dep, reg, DATE, outcomeDate, outcome, Population and hosp variables.}

\item{param_Y}{The param_Y hyperparameter equal to 'hosp' for raw prediction or 'HOSPHAB' for per 100 000 inhabitants prediction}

\item{param_VARIATION}{The param_VARIATION hyperparameter, taking into account raw prediction or variation prediction}

\item{s}{The s to select the right model from cv.glmnet}
}
\value{
A dataframe with the predictions and dep, reg, DATE, outcomeDate, outcome, Population and hosp columns from fullTest
}
\description{
Compute predictions for glmnet taking into account param_VARIATION and param_Y for prediction.
}
