% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctQuantPredictionRanger.R
\name{FctQuantPredictionRanger}
\alias{FctQuantPredictionRanger}
\title{FctQuantPredictionRanger}
\usage{
FctQuantPredictionRanger(
  rangerModel,
  qinf = 0.025,
  qsup = 0.975,
  dfTestX,
  dfTestFull,
  param_Y,
  param_VARIATION
)
}
\arguments{
\item{rangerModel}{A ranger random forest model}

\item{qinf}{quantile inf (default = 0.025)}

\item{qsup}{quantile sup (default = 0.975)}

\item{dfTestX}{The X test}

\item{dfTestFull}{The full tests set for convenient output}

\item{param_Y}{the Y hyperparam}

\item{param_VARIATION}{The VARIATION hyperparam}
}
\value{
A dataframe with predictions
}
\description{
Quantile prediction from ranger dealing with param_VARIATION and HOSPHAB hyperparam
}
