% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctSelectFeatures.R
\name{FctSelectFeatures}
\alias{FctSelectFeatures}
\title{FctSelectFeatures}
\usage{
FctSelectFeatures(df, param_X)
}
\arguments{
\item{df}{A dataframe with all the variables}

\item{param_X}{A character vector either: 'hosp', 'pcr', or 'pcrhosp'}
}
\value{
Remove the selected columns from the dataframe
}
\description{
Function to select relevant variables as mentionned by param_X
}
