% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctTrainRangerQRF.R
\name{FctTrainRangerQRF}
\alias{FctTrainRangerQRF}
\title{FctTrainRangerQRF}
\usage{
FctTrainRangerQRF(
  yTrain,
  xTrain,
  ntree,
  xTest,
  fullTest,
  varMarginalEffect = NULL,
  param_Y,
  param_X,
  param_DATE,
  param_FORECAST,
  param_VARIATION
)
}
\arguments{
\item{yTrain}{The y train}

\item{xTrain}{The x train dataset}

\item{ntree}{the number of trees}

\item{xTest}{the x test dataframe}

\item{fullTest}{the full x test for nice output}

\item{varMarginalEffect}{Vector of variable for which marginal effect will be evaluated on the train set.}

\item{param_Y}{the Y hyperparam}

\item{param_X}{the X hyperparam}

\item{param_DATE}{the date hyperparam}

\item{param_FORECAST}{the forecast hyperparam}

\item{param_VARIATION}{the variation hyperparam}
}
\value{
a list of 2: variable importance and the predictions
}
\description{
Function to compute ranger random forest
}
