% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctTrainSvm.R
\name{FctTrainSvm}
\alias{FctTrainSvm}
\title{FctTrainSvm}
\usage{
FctTrainSvm(
  yTrain,
  xTrain,
  xTest,
  fullTest,
  param_Y,
  param_X,
  param_DATE,
  param_FORECAST,
  param_VARIATION,
  tuneCost = 10,
  tuneGamma = 0.1,
  tumeEpsilon = 10^(-2),
  kernel = "radial"
)
}
\arguments{
\item{yTrain}{The y vector used for training}

\item{xTrain}{The x dataframe used for training}

\item{xTest}{The x dataframe used for testing}

\item{fullTest}{The full dataset for testing and reporting}

\item{param_Y}{The Y hyperparam ('hosp' or 'HOSPHAB')}

\item{param_X}{The X hyperparam ('hosp', 'pcr' or 'pcrhosp')}

\item{param_DATE}{The DATE hyperparam (a date)}

\item{param_FORECAST}{The forecast hyperparam (a double)}

\item{param_VARIATION}{The variation hyperparam (a boolean, if true, learning is done on the variation instead of raw outcome)}

\item{tuneCost}{Numeric vector of tuning values for cost parameter. Default is 10^c(0:2).}

\item{tuneGamma}{Numeric vector of tuning values for gamma parameter. Default is c(0.1, 0.2, 0.5).}

\item{tumeEpsilon}{Numeric vector of tuning values for epsilon parameter. Default is 10^c(0:2).}

\item{kernel}{The kernel type. Default is 'radial'.}
}
\value{
A list of two dataframe: the variable importance and the predictions
}
\description{
Covid prediction with SVM model.
}
