% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctTrainTestSidepDelay.R
\name{FctTrainTestSidepDelay}
\alias{FctTrainTestSidepDelay}
\title{FctTrainTestSidepDelay}
\usage{
FctTrainTestSidepDelay(df, param_DATE, delaySIDEP, param_VARIATION)
}
\arguments{
\item{df}{Dataframe of interest}

\item{param_DATE}{The date of interest}

\item{delaySIDEP}{The SIDEP delay compared to CHU}

\item{param_VARIATION}{A boolean, should the model be trained on raw value or on the delta}
}
\value{
A list with train and test sets
}
\description{
Function to generate train and test sets
}
