% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctVarImpDfPlot.R
\name{FctVarImpDfPlot}
\alias{FctVarImpDfPlot}
\title{FctVarImpDfPlot}
\usage{
FctVarImpDfPlot(
  dfImp,
  Features = "Features",
  Importance = "Importance",
  groupVar = "param_FORECAST",
  topN = 20,
  CleanFeaturesName = TRUE
)
}
\arguments{
\item{dfImp}{A dataframe with all columns needed as specified by following arguments}

\item{Features}{Features column name}

\item{Importance}{Importance column name}

\item{groupVar}{Grouping variables column name}

\item{topN}{Only importance for the topN variables according to median will be computed (default is 20)}

\item{CleanFeaturesName}{A boolean (default is TRUE). If true, features names are passed through FctCleanFeaturesName() fonction to clean the features names.}
}
\value{
A dataframe with the columns as specified in Feautres, Importance and groupVar for the topN variables
}
\description{
Compute the optimal dataframe to plot variables importance
}
