% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctVarImpEnet.R
\name{FctVarImpEnet}
\alias{FctVarImpEnet}
\title{FctVarImpEnet}
\usage{
FctVarImpEnet(modelEnet, s = "lambda.1se")
}
\arguments{
\item{modelEnet}{An elastic net model from cv.glmnet() function}

\item{s}{The s used to choose the right beta values}
}
\value{
The dataframe for a given s with two columns: (i) the Features and (ii) the beta.
}
\description{
FctVarImpEnet
}
