% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FctVarImpHyperparam.R
\name{FctVarImpHyperparam}
\alias{FctVarImpHyperparam}
\title{FctVarImpHyperparam}
\usage{
FctVarImpHyperparam(rangerModel)
}
\arguments{
\item{rangerModel}{ranger random forest model}
}
\value{
A dataframe with variable importance
}
\description{
Clean varimp from ranger model
}
