% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSelectPenalised.R
\name{FeatSelectPenalised}
\alias{FeatSelectPenalised}
\title{FeatSelectPenalised}
\usage{
FeatSelectPenalised(df, param_FORECAST = 7, alpha = 1, nboot = 100)
}
\arguments{
\item{df}{The dataframe of interest}

\item{param_FORECAST}{The forecast length.}

\item{alpha}{The penalisation constraint as in glmnet (default is 1, lasso penalisation).}

\item{nboot}{Number of bootstrap samples the model should be trained (for robustness).}
}
\value{
A a vetor with the non zero variables.
}
\description{
[Production] Train a lasso linear regression model and give the selected variables as output depending on the forecast time of interest.
}
