% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestModel.R
\name{FindBestModel}
\alias{FindBestModel}
\title{FindBestModel}
\usage{
FindBestModel(
  dfPerf,
  forecast = 7,
  vecOutcome = c("Prevalent variation", "HOSP_DERIV", "Prevalent")
)
}
\arguments{
\item{dfPerf}{The performance dataframe}

\item{forecast}{The forecast time}

\item{vecOutcome}{The vector of outcome considered}
}
\value{
The label of the best model
}
\description{
Find the best model for a given forecast time
}
