% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestModelbyOutcome.R
\name{FindBestModelbyOutcome}
\alias{FindBestModelbyOutcome}
\title{FindBestModelbyOutcome}
\usage{
FindBestModelbyOutcome(dfPerf)
}
\arguments{
\item{dfPerf}{The performance dataframe from Dfperformances().}
}
\value{
A dataframe with the best model by outcome
}
\description{
Find the best model by outcome
}
