% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastLm.R
\name{ForecastLm}
\alias{ForecastLm}
\title{ForecastLm}
\usage{
ForecastLm(
  df,
  variableSubset = c("hosp", "hosp_fct_firstDeriv7", "dep",
    "hosp_fct_firstDeriv7_fct_firstDeriv7", "P_tous_ages", "P_tous_ages_fct_firstDeriv7",
    "P_tous_ages_fct_firstDeriv7_fct_firstDeriv7", "FracP_tous_ages_fct_firstDeriv7",
    "FracP_tous_ages_fct_firstDeriv7_fct_firstDeriv7", "FracP_tous_ages"),
  param_FORECAST = 7,
  interactionPCR = F,
  param_DATE,
  delaySIDEP = 4
)
}
\arguments{
\item{df}{The dataframe of interest}

\item{variableSubset}{The subset of variables.}

\item{param_FORECAST}{The forecast length.}

\item{interactionPCR}{Boolean. Should an interaction between dep and P_tous_ages be added.}

\item{param_DATE}{The date from which forecast is done.}

\item{delaySIDEP}{Delay of SIDEP data (default is 4 for historical analysis)}
}
\value{
A list of two dataframe with the Prediction and the model coefficients
}
\description{
[Production] Train linear regression model on the subset of variable specified and return forecasting at given time.
}
\examples{
\dontrun{
library(PredictCovid)
##################### import datasets ##################
## CHU file
load(file = "extdata/COVID_ferte/dataCovidCHU.rdata")
## SIDEP + DM
dfSidepChu <- ImportCleanSIDEPCHU(df_chu = dataCovidCHU)

##################### train elastic-net ##################
vecForecast <- 1:14
dfForecast <- sapply(vecForecast,
                     simplify = F,
                     function(param_FORECAST) ForecastEnet(df = dfSidepChu,
                                                           param_FORECAST = param_FORECAST))

dfPredObs <- bind_rows(dfForecast) \%>\%
         rename("DATE" = "outcomeDate",
                "hosp" = "Prediction") \%>\%
         mutate(STATUS = "Prediction") \%>\%
           bind_rows(dfSidepChu \%>\%
                         select(dep, DATE, hosp) \%>\%
                          mutate(STATUS = "Observed"))

}
}
