% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastLmTime.R
\name{ForecastLmTime}
\alias{ForecastLmTime}
\title{ForecastLmTime}
\usage{
ForecastLmTime(Forecast, vecSidepDate, dfSidep)
}
\arguments{
\item{Forecast}{The forecast time}

\item{vecSidepDate}{The vector of dates}

\item{dfSidep}{The dataframe of prediction}
}
\value{
A dataframe with the predictions.
}
\description{
Function to compute ForecastLm on a vector of Date.
}
