% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromDfToArray.R
\name{FromDfToArray}
\alias{FromDfToArray}
\title{FromDfToArray}
\usage{
FromDfToArray(df)
}
\arguments{
\item{df}{A dataframe with tume and id columns}
}
\value{
An array for Julia Frechet
}
\description{
Transform dataframe to array
}
