% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_weighted_daily_obs.R
\name{fun_weighted_daily_obs}
\alias{fun_weighted_daily_obs}
\title{fun_weighted_daily_obs}
\usage{
fun_weighted_daily_obs(data0, which_insee, daily_station_weather)
}
\arguments{
\item{data0}{A dataframe with the following columns: code_insee (the insee code), nom (the insee area label), station (the meteorological station), code (the station code), total_pop_buff (the population inside the 10km buffer around the station)}

\item{which_insee}{The insee code evaluated.}

\item{daily_station_weather}{The meteorological data set with the following columns: code (the station code), date_day (the date), stat_ (multiples functions with the meteorological data collected).}
}
\value{
A dataframe with 9 columns:
\itemize{
  \item label_insee - The name of the region or the departement.
  \item code_insee - The insee code of the departement or the region.
  \item date_day - The date.
  \item stat_t.mean - The mean temperature over the day..
  \item stat_precip - The precipitation over the day.
  \item stat_RH.mean - The mean relative humidity over the day.
  \item stat_AH.mean - The mean absolute humidity over the day.
  \item stat_IPTCC.mean - The mean IPTCC index over the day.
  \item stat_ws.mean - The mean wind speed over the day.
  \item stat_dewpoint.mean - The mean dew point over the day.
}
}
\description{
Function giving daily weighted average weather data for a given area (INSEE code)
}
\examples{

data0 <- data.frame(code_insee = c(1, 1, 2),
                   nom = c("Tatooine", "Tatooine", "Alderaan"),
                   station = c("Mos Eisley", "Mos Espa", "Alderaan city"),
                   code = c(118, 218, 104),
                   total_pop_buff = c(0.6, 0.4, 1.0))
 daily_station_weather <- data.frame(
 code = rep(c(118, 218, 104), 2),
 date_day = lubridate::as_date(x = c(rep("2080-01-01", 3),
                                     rep("2080-01-02", 3))),
 stat_t.mean = c(32, 34, 18, 36, 35, 15),
 stat_precip = c(32, 34, 18, 36, 35, 15),
 stat_RH.mean = c(32, 34, 18, 36, 35, 15),
 stat_AH.mean = c(32, 34, 18, 36, 35, 15),
 stat_IPTCC.mean = c(32, 34, 18, 36, 35, 15),
 stat_ws.mean = c(32, 34, 18, 36, 35, 15),
 stat_dewpoint.mean = c(32, 34, 18, 36, 35, 15)
 )

which_insee <- 1
fun_weighted_daily_obs(data0 = data0,
                       which_insee = which_insee,
                       daily_station_weather = daily_station_weather)

}
