% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateGgplotLabels.R
\name{GenerateGgplotLabels}
\alias{GenerateGgplotLabels}
\title{GenerateGgplotLabels}
\usage{
GenerateGgplotLabels(vecLabels)
}
\arguments{
\item{vecLabels}{A vector of labels to be cleaned}
}
\value{
A cleaned named vector of labels
}
\description{
Clean labels for ggplot graph
}
