% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotBestFeatures.R
\name{ggplotBestFeatures}
\alias{ggplotBestFeatures}
\title{ggplotBestFeatures}
\usage{
ggplotBestFeatures(
  bestModelCompromise,
  dflsDataGroupImp,
  dflsDataImp,
  dfDashboardRaw
)
}
\arguments{
\item{bestModelCompromise}{The best model}

\item{dflsDataGroupImp}{The translator of ID and model characteristics}

\item{dflsDataImp}{The variable importance}

\item{dfDashboardRaw}{The raw data}
}
\value{
A ggplot object
}
\description{
Plot the best features of the best model
}
