% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotBestModel.R
\name{ggplotBestModel}
\alias{ggplotBestModel}
\title{ggplotBestModel}
\usage{
ggplotBestModel(df, dfDashboard)
}
\arguments{
\item{df}{A dataframe with model predictions}

\item{dfDashboard}{The raw data}
}
\value{
A ggplot object
}
\description{
Plot a model predictions in the live prediction way.
}
