% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotBestModelError.R
\name{ggplotBestModelError}
\alias{ggplotBestModelError}
\title{ggplotBestModelError}
\usage{
ggplotBestModelError(df)
}
\arguments{
\item{df}{The best model dataframe}
}
\value{
A list with month error dataframe and plot
}
\description{
Plot best model with mean relative and absolute error
}
