% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotBestPredictions.R
\name{ggplotBestPredictions}
\alias{ggplotBestPredictions}
\title{ggplotBestPredictions}
\usage{
ggplotBestPredictions(dfPredictionsBest)
}
\arguments{
\item{dfPredictionsBest}{Prediction dataframe with outcome, CONSTANT and Prediction columns}
}
\value{
A ggplot object
}
\description{
Plot the predictions of a model with constant prediction for comparison.
}
