% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotBestvsOthers.R
\name{ggplotBestvsOthers}
\alias{ggplotBestvsOthers}
\title{ggplotBestvsOthers}
\usage{
ggplotBestvsOthers(df, dfNPI)
}
\arguments{
\item{df}{Prediction dataframe}

\item{dfNPI}{NPI dataframe}
}
\value{
A ggplot object
}
\description{
Plot model predictions depending on NPI
}
