% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotErrorModelComparison.R
\name{ggplotErrorModelComparison}
\alias{ggplotErrorModelComparison}
\title{ggplotErrorModelComparison}
\usage{
ggplotErrorModelComparison(
  dfResultsMergedPred,
  dfSidepLmPred,
  trainTestDate,
  dfDashboard,
  dfPredresultSensitivitySidep,
  type
)
}
\arguments{
\item{dfResultsMergedPred}{The dataframe from which the best prediction are extracted}

\item{dfSidepLmPred}{The dataframe with sidep model predictions}

\item{trainTestDate}{The train test date}

\item{dfDashboard}{The raw data}

\item{dfPredresultSensitivitySidep}{The sensitivity analysis removing sidep data}

\item{type}{The type either "SIDEP" or "NoSidep"}
}
\value{
A ggplot object
}
\description{
Plot the difference of prediction of best model compared to other models
}
