% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotFeaturesManualSet.R
\name{ggplotFeaturesManualSet}
\alias{ggplotFeaturesManualSet}
\title{ggplotFeaturesManualSet}
\usage{
ggplotFeaturesManualSet(dfDashboard, vecFeatures)
}
\arguments{
\item{dfDashboard}{The data}

\item{vecFeatures}{The vector of features}
}
\value{
A ggplot object
}
\description{
Plot features of a manual selection
}
