% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotFigImp.R
\name{ggplotFigImp}
\alias{ggplotFigImp}
\title{ggplotFigImp}
\usage{
ggplotFigImp(
  bestModelCompromise,
  dflsDataGroupImp,
  dflsDataImp,
  ModelType = "Lasso",
  nbFeatures = 20
)
}
\arguments{
\item{bestModelCompromise}{The best model}

\item{dflsDataGroupImp}{The translator between ID and model characteristics.}

\item{dflsDataImp}{The importance}

\item{ModelType}{The model either "Lasso" or "Rf - Permutation"}

\item{nbFeatures}{The nb of features to be displayed}
}
\value{
A ggplot object
}
\description{
Show the feature importance of the best model
}
