% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotFigImpLasso.R
\name{ggplotFigImpLasso}
\alias{ggplotFigImpLasso}
\title{ggplotFigImpLasso}
\usage{
ggplotFigImpLasso(dfPredictionsBest, dflsDataGroupImp, dflsDataImp)
}
\arguments{
\item{dfPredictionsBest}{Best model predictions.}

\item{dflsDataGroupImp}{The translation between id and characteristics.}

\item{dflsDataImp}{The data to import.}
}
\value{
A ggplot object
}
\description{
Show the lasso importance of the best model.
}
