% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotModelbyForecast.R
\name{ggplotModelbyForecast}
\alias{ggplotModelbyForecast}
\title{ggplotModelbyForecast}
\usage{
ggplotModelbyForecast(df, dfDashboard, margin = FALSE, trainTestDate = NULL)
}
\arguments{
\item{df}{A dataframe with predictions}

\item{dfDashboard}{A raw dataframe}

\item{margin}{Should 10 and 20\% margin be shown}

\item{trainTestDate}{Inferior limit of plot default is NULL}
}
\value{
A ggplot object
}
\description{
Prediction by forecast time
}
