% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMultipleLag.R
\name{ggplotMultipleLag}
\alias{ggplotMultipleLag}
\title{ggplotMultipleLag}
\usage{
ggplotMultipleLag(vecMultipleLag, dfDashboardRaw, dfPredictions)
}
\arguments{
\item{vecMultipleLag}{The vector of multiple lag (maximum must be the forecast time)}

\item{dfDashboardRaw}{The raw features}

\item{dfPredictions}{The predictions dataframe}
}
\value{
A ggplot object
}
\description{
Plot multiple lag predictions
}
