% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotNPI.R
\name{ggplotNPI}
\alias{ggplotNPI}
\title{ggplotNPI}
\usage{
ggplotNPI(dfDashboard)
}
\arguments{
\item{dfDashboard}{data}
}
\value{
A plot of NPI, IPTCC and hospitalisations
}
\description{
Generate NPI plot from raw data
}
