% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPerf.R
\name{ggplotPerf}
\alias{ggplotPerf}
\title{ggplotPerf}
\usage{
ggplotPerf(dfPerf, type = "Global")
}
\arguments{
\item{dfPerf}{The performance dataframe from Dfperformances().}

\item{type}{Should either "Global", "Simplified" or "SimplifiedNoSpan".}
}
\value{
A ggplot object
}
\description{
Plot performance
}
