% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPredVariation.R
\name{ggplotPredVariation}
\alias{ggplotPredVariation}
\title{ggplotPredVariation}
\usage{
ggplotPredVariation(dfPredictionsVariationBest, vecLabelsGG)
}
\arguments{
\item{dfPredictionsVariationBest}{dataframe of best model predictions}

\item{vecLabelsGG}{Vector of labels}
}
\value{
A ggplot object
}
\description{
Prediction variation of the best model
}
