% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotURGENCES.R
\name{ggplotURGENCES}
\alias{ggplotURGENCES}
\title{ggplotURGENCES}
\usage{
ggplotURGENCES(dfDashboard)
}
\arguments{
\item{dfDashboard}{Data}
}
\value{
A ggplot object
}
\description{
Generate plot of emergency units data
}
