% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotURGENCESSA.R
\name{ggplotURGENCESSA}
\alias{ggplotURGENCESSA}
\title{ggplotURGENCESSA}
\usage{
ggplotURGENCESSA(dfDashboard)
}
\arguments{
\item{dfDashboard}{Data}
}
\value{
A ggplot object
}
\description{
Generate plot of emergency units data
}
