% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotVaccinVariants.R
\name{ggplotVaccinVariants}
\alias{ggplotVaccinVariants}
\title{ggplotVaccinVariants}
\usage{
ggplotVaccinVariants(dfDashboard)
}
\arguments{
\item{dfDashboard}{The raw data}
}
\value{
a ggplot object
}
\description{
Generate plot of vaccin and variants
}
