% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupDashboardPCRAge.R
\name{GroupDashboardPCRAge}
\alias{GroupDashboardPCRAge}
\title{GroupDashboardPCRAge}
\usage{
GroupDashboardPCRAge(df)
}
\arguments{
\item{df}{The dashboard dataframe}
}
\value{
A dataframe with the grouped categories
}
\description{
Groupe PCR age categories in 0-19, 20-59 and 60-more
}
