% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportCleanSIDEPCHU.R
\name{ImportCleanSIDEPCHU}
\alias{ImportCleanSIDEPCHU}
\title{ImportCleanSIDEPCHU}
\usage{
ImportCleanSIDEPCHU(df_chu)
}
\arguments{
\item{df_chu}{The CHU dataframe.}
}
\value{
The datamanaged dataframe
}
\description{
Import SIDEP data, aggregate with CHU and datamanage it.
}
\examples{
\dontrun{
## CHU file
load(file = "extdata/COVID_ferte/dataCovidCHU.rdata")

ImportCleanSIDEPCHU(df_chu = dataCovidCHU)

}
}
