% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportCleanVaccination.R
\name{ImportCleanVaccination}
\alias{ImportCleanVaccination}
\title{ImportCleanVaccination}
\usage{
ImportCleanVaccination(
 
    urlVaccin = "https://www.data.gouv.fr/en/datasets/r/83cbbdb9-23cb-455e-8231-69fc25d58111"
)
}
\arguments{
\item{urlVaccin}{the vaccin data url}
}
\value{
A data frame with departement, date and vaccin cumulative doses by age class
}
\description{
Import vaccin data by department and datamange it
}
