% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportGompertzImputeVariants.R
\name{ImportGompertzImputeVariants}
\alias{ImportGompertzImputeVariants}
\title{ImportGompertzImputeVariants}
\usage{
ImportGompertzImputeVariants(
 
    fileVariantsBefore062021 = "https://www.data.gouv.fr/en/datasets/r/16f4fd03-797f-4616-bca9-78ff212d06e8",
 
    fileVariantsAfter062021 = "https://www.data.gouv.fr/fr/datasets/r/4d3e5a8b-9649-4c41-86ec-5420eb6b530c",
  datePlateauVariants = as.Date("2021-04-15"),
  dfVariantsImputed = NULL
)
}
\arguments{
\item{fileVariantsBefore062021}{the url of the variants before 06 2021.}

\item{fileVariantsAfter062021}{the url of the variants after 06 2021.}

\item{datePlateauVariants}{the max date of the data used to train Gompertz. This is necessary because of the decrease in april.}

\item{dfVariantsImputed}{The previous dataset with imputation. If not provided, the Gompertz imputation is performed for missing values.}
}
\value{
A dataframe with department, date, variants percentage
}
\description{
Import Variants df and impute missing data with Gompertz law
}
