% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMajorityVariant.R
\name{ImportMajorityVariant}
\alias{ImportMajorityVariant}
\title{ImportMajorityVariant}
\usage{
ImportMajorityVariant(
 
    fileVariantsBefore062021 = "https://www.data.gouv.fr/en/datasets/r/16f4fd03-797f-4616-bca9-78ff212d06e8",
 
    fileVariantsAfter062021 = "https://www.data.gouv.fr/fr/datasets/r/4d3e5a8b-9649-4c41-86ec-5420eb6b530c",
 
    fileVariantsAfter012022 = "https://www.data.gouv.fr/fr/datasets/r/ba8219dc-948c-418a-9116-f792c79c54b8"
)
}
\arguments{
\item{fileVariantsBefore062021}{the url of the variants before 06 2021.}

\item{fileVariantsAfter062021}{the url of the variants after 06 2021.}

\item{fileVariantsAfter012022}{the url of the variants after 01 2022.}
}
\value{
A dataframe with the date, the department and the majority variant.
}
\description{
Import Variants df and find the majority variant at a given date by department.
}
