% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportOpenData.R
\name{ImportOpenData}
\alias{ImportOpenData}
\title{ImportOpenData}
\usage{
ImportOpenData()
}
\value{
A list of length two with the merged dataframe and the delay between datasources
}
\description{
Import data from covid openscience sources (vaccination, variants, weather, hospitalisation, pcr by age)
}
