% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeMissingWeather.R
\name{ImputeMissingWeather}
\alias{ImputeMissingWeather}
\title{ImputeMissingWeather}
\usage{
ImputeMissingWeather(df_to_imp)
}
\arguments{
\item{df_to_imp}{Dataframe to impute.}
}
\value{
A list of three : the dataframe of interest and to diagnosis dataframe (before mean adjacent step and before LOCF step)
}
\description{
Imputation is first done by providing the mean of the adjacent departement and then using the LOCF.
}
