% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassographicalCheck.R
\name{LassographicalCheck}
\alias{LassographicalCheck}
\title{LassographicalCheck}
\usage{
LassographicalCheck(df, param_FORECAST = 7, alpha = 1)
}
\arguments{
\item{df}{The dataframe of interest}

\item{param_FORECAST}{The forecast length.}

\item{alpha}{The penalisation constraint as in glmnet (default is 1, lasso penalisation).}
}
\value{
A vector with the non zero variables.
}
\description{
Train a lasso linear regression model and give the selected variables as output depending on the forecast time of interest.
}
