% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadVariants.R
\name{LoadVariants}
\alias{LoadVariants}
\title{LoadVariants}
\usage{
LoadVariants(
  fileVariantsBefore062021,
  fileVariantsAfter062021,
  fileVariantsAfter012022,
  addVariantsBefore062021 = TRUE
)
}
\arguments{
\item{fileVariantsBefore062021}{the url of the variants before 06 2021.}

\item{fileVariantsAfter062021}{the url of the variants after 06 2021.}

\item{fileVariantsAfter012022}{the url of the variants after 01 2022.}

\item{addVariantsBefore062021}{should the variants before 2021-06 be added. Deprecated.}
}
\value{
A dataframe with department, date, variants percentage
}
\description{
Import Variants df
}
