% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsfct_meanCasesDeriv.R
\name{ls_fctmeanCasesDeriv}
\alias{ls_fctmeanCasesDeriv}
\title{ls_fctmeanCasesDeriv}
\usage{
ls_fctmeanCasesDeriv(vec_lagTimepoints = c(1, 3, 7, 14))
}
\arguments{
\item{vec_lagTimepoints}{A vector of lag time points (default is c(1, 3, 7, 14))}
}
\value{
A list of function for past cumulating mean Cases and first derivative.
}
\description{
Build a list of function to get first derivative and mean number of cases in the past vec_lagTimepoints.
}
