% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianIQR.R
\name{MedianIQR}
\alias{MedianIQR}
\title{MedianIQR}
\usage{
MedianIQR(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{remove NA (default is TRUE)}
}
\value{
A character string with median and IQR
}
\description{
Compute the median and the IQR, return a nice text
}
