% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeDashboardVaccinVariantsWeather.R
\name{MergeDashboardVaccinVariantsWeather}
\alias{MergeDashboardVaccinVariantsWeather}
\title{MergeDashboardVaccinVariantsWeather}
\usage{
MergeDashboardVaccinVariantsWeather(
  dfFullDashboard,
  dfVariants_dep33,
  dfVaccination_dep33,
  weather_data_dep33 = weather_data_dep33,
  vecExcludeTransf = c("START_DATE", "n_cum_dose1_tous_ages", "n_cum_dose1_18_59",
    "n_cum_dose1_60_80_PLUS", "Pct_Variants")
)
}
\arguments{
\item{dfFullDashboard}{The Dashboard data}

\item{dfVariants_dep33}{The variants data}

\item{dfVaccination_dep33}{The vaccine data}

\item{weather_data_dep33}{The weather data}

\item{vecExcludeTransf}{Vector of features for which feature engineering should not be performed}
}
\value{
A list of dataframe
}
\description{
Merge Dashboard, Vaccin, Variants and Weather
}
