% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeFolderPredictions.R
\name{MergeFolderPredictions}
\alias{MergeFolderPredictions}
\title{MergeFolderPredictions}
\usage{
MergeFolderPredictions(folder)
}
\arguments{
\item{folder}{The folder path}
}
\value{
A list of predictions and features importance
}
\description{
Merge folder PredictCovid Data
}
