% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeOpenData.R
\name{MergeOpenData}
\alias{MergeOpenData}
\title{MergeOpenData}
\usage{
MergeOpenData(dfVariants, dfVaccination, weather_data_byDep, df_hospitPCR)
}
\arguments{
\item{dfVariants}{Variant data frame}

\item{dfVaccination}{Vaccine data frame}

\item{weather_data_byDep}{Weather data frame}

\item{df_hospitPCR}{Hospitalisation and PCR data frame}
}
\value{
A list of length two with the merged dataframe and the delay between datasources
}
\description{
Merge open data (vaccination, variants, weather, hospitalisation, pcr by age)
}
