% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeRfandEnetPredictions.R
\name{MergeRfandEnetPredictions}
\alias{MergeRfandEnetPredictions}
\title{MergeRfandEnetPredictions}
\usage{
MergeRfandEnetPredictions(df_test, lsmodelEnet, lsmodelRf, OUTCOME, FORECAST)
}
\arguments{
\item{df_test}{The test dataframe}

\item{lsmodelEnet}{The list from trainElasticNet()}

\item{lsmodelRf}{The list from trainRf()}

\item{OUTCOME}{The OUTCOME must be one of HOSP or IN_HOSP}

\item{FORECAST}{The FORECAST time parameter}
}
\value{
The predictions df
}
\description{
Merge Elastic net and random forest predictions
}
