% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleLag.R
\name{MultipleLag}
\alias{MultipleLag}
\title{MultipleLag}
\usage{
MultipleLag(vecNum = 1:14)
}
\arguments{
\item{vecNum}{The numeric vector corresponging to the different lag.}
}
\value{
A list of functions
}
\description{
Create multiple lag functions with nice names.
}
