% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStepFeatureSelection.R
\name{OneStepFeatureSelection}
\alias{OneStepFeatureSelection}
\title{OneStepFeatureSelection}
\usage{
OneStepFeatureSelection(OUTCOME, FORECAST, df)
}
\arguments{
\item{OUTCOME}{The outcome either HOSP or IN_HOSP}

\item{FORECAST}{The forecast time}

\item{df}{The dataframe of interest assuming date is stored in START_DATE column}
}
\value{
A dataframe with variable importance of RF and Lasso
}
\description{
Function to compute variable importance from outcome and forecast hyperparameters using lasso and random forest
}
