% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputLme4.R
\name{OutputLme4}
\alias{OutputLme4}
\title{OutputLme4}
\usage{
OutputLme4(lme4Object)
}
\arguments{
\item{lme4Object}{Lme4 object}
}
\value{
A dataframe with coef and CI
}
\description{
Nice output for lme4 objects.
}
