% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerfFromdfPredObs.R
\name{PerfFromdfPredObs}
\alias{PerfFromdfPredObs}
\title{PerfFromdfPredObs}
\usage{
PerfFromdfPredObs(dfPredObs)
}
\arguments{
\item{dfPredObs}{The dataframe to be evaluated}
}
\value{
a dataframe with departement, forecast time, period of evaluation, absolute and relative errors and coverage percentage.
}
\description{
Compute the perf table from the dfPredObs obtained with ForecastLm function.
}
