% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonLmReFe.R
\name{PoissonLmReFe}
\alias{PoissonLmReFe}
\title{PoissonLmReFe}
\usage{
PoissonLmReFe(df, param_FORECAST, param_DATE, featureSelection)
}
\arguments{
\item{df}{Data}

\item{param_FORECAST}{The forecast time}

\item{param_DATE}{The date from which prediction is made}

\item{featureSelection}{The variables used for prediction}
}
\value{
A list of three dataframes (Coefficients, predictions and marginal department effect)
}
\description{
Run Poisson and lm with and with dep fixed effect or dep/reg random effect.
}
