% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictFrechet.R
\name{PredictFrechet}
\alias{PredictFrechet}
\title{PredictFrechet}
\usage{
PredictFrechet(
  dfOutcomeEngineered,
  dbl_curvePastWindow = 14,
  date,
  ntree = 100,
  ncores = 1,
  boot = FALSE
)
}
\arguments{
\item{dfOutcomeEngineered}{The dataframe}

\item{dbl_curvePastWindow}{The frechet window}

\item{date}{The date of test set}

\item{ntree}{Nb of frechet tree to build}

\item{ncores}{Nb of cores to use}

\item{boot}{should bootstrap sampling be performed on train set}
}
\value{
A list of two dataframe: predictions and feature importance
}
\description{
Make Frechet random forest prediction and get the prediction dataframe (variable importance is not supported)
}
