% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionDashboardFromHyperparam.R
\name{PredictionDashboardFromHyperparam}
\alias{PredictionDashboardFromHyperparam}
\title{PredictionDashboardFromHyperparam}
\usage{
PredictionDashboardFromHyperparam(
  df,
  lsFeatures,
  SELECTION,
  OUTCOME,
  DATE,
  FORECAST
)
}
\arguments{
\item{df}{The dataframe}

\item{lsFeatures}{The list of features computed by the hyperparam_chu_dashboard.R script}

\item{SELECTION}{The selection parameter muste be one of the following: VI (user input), VI_7dMaxDeriv (another user input), vecVI_small (a third user input), All (all features are used), One_step (selection is done prior to 10-12-2020), Adaptative (selection is done using lasso and rf, for rf a number of features equals to lasso length of selection is used), Adaptative_3months (same as adaptative but the training of this model is used only on previous 3 months)}

\item{OUTCOME}{The OUTCOME must be one of HOSP or IN_HOSP}

\item{DATE}{The date from which prediction is made}

\item{FORECAST}{The forecast time.}
}
\value{
A list of two dataframes : dfPred (predictions) and dfImportance (variable importance)
}
\description{
Compute predictions for CHU dashboard model depending on hyperparameters.
}
