% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredVarimpFromFolder.R
\name{PredVarimpFromFolder}
\alias{PredVarimpFromFolder}
\title{PredVarimpFromFolder}
\usage{
PredVarimpFromFolder(folder)
}
\arguments{
\item{folder}{The folder path}
}
\value{
A list of length 2 : predictions and variable importance dataframe
}
\description{
Function to extract prediction and variable importance from a folder.
}
