% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preplot_sumOverFrRegDep.R
\name{preplot_sumOverFrRegDep}
\alias{preplot_sumOverFrRegDep}
\title{preplot_sumOverFrRegDep}
\usage{
preplot_sumOverFrRegDep(
  df,
  char_colDate,
  vec_colMeasure,
  char_region = "reg",
  char_dep = "dep"
)
}
\arguments{
\item{df}{Dataframe}

\item{char_colDate}{The date column date}

\item{vec_colMeasure}{The vector of measure of interest column names}

\item{char_region}{The region column name}

\item{char_dep}{The departement column name}
}
\value{
A dataframe with measure aggregated at Country, region and departement by date and by departement.
}
\description{
Compute a preplot dataframe summing a value over a Country, region, departement by date.
}
