% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_orderboxplot.R
\name{reorder_within}
\alias{reorder_within}
\title{reorder_within}
\usage{
reorder_within(x, by, within, fun = mean, sep = "___", ...)
}
\arguments{
\item{x}{internal}

\item{by}{internal}

\item{within}{internal}

\item{fun}{internal}

\item{sep}{internal}

\item{...}{internal}
}
\value{
internal
}
\description{
function to order boxplot output
}
