% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolFunctions.R
\name{rolDeriv}
\alias{rolDeriv}
\title{rolDeriv}
\usage{
rolDeriv(x, lagDays)
}
\arguments{
\item{x}{feature}

\item{lagDays}{The deepness of the derivative}
}
\value{
The transformed feature
}
\description{
Deriv over last lagDays days
}
