% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_orderboxplot.R
\name{scale_y_reordered}
\alias{scale_y_reordered}
\title{scale_y_reordered}
\usage{
scale_y_reordered(..., sep = "___")
}
\arguments{
\item{...}{internal}

\item{sep}{internal}
}
\value{
internal
}
\description{
function to order boxplot output
}
