% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparateTrainTest.R
\name{SeparateTrainTest}
\alias{SeparateTrainTest}
\title{SeparateTrainTest}
\usage{
SeparateTrainTest(df, DATE, SELECTION, SKIP_CLUSTER = NULL)
}
\arguments{
\item{df}{The dataframe}

\item{DATE}{The date of split}

\item{SELECTION}{The selection parameter muste be one of the following: VI (user input), VI_7dMaxDeriv (another user input), vecVI_small (a third user input), All (all features are used), One_step (selection is done prior to 10-12-2020), Adaptative (selection is done using lasso and rf, for rf a number of features equals to lasso length of selection is used), Adaptative_3months (same as adaptative but the training of this model is used only on previous 3 months)}

\item{SKIP_CLUSTER}{A named list of two elements : the "start" and "end" dates of the cluster}
}
\value{
A list with train and test sets
}
\description{
Separate train and test data
}
