% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothDf.R
\name{smoothDf}
\alias{smoothDf}
\title{smoothDf}
\usage{
smoothDf(
  df,
  DATE_colum = "START_DATE",
  span_days = 14,
  skip_variables = NULL,
  degree = 1
)
}
\arguments{
\item{df}{The dataframe}

\item{DATE_colum}{The date column name}

\item{span_days}{The span (default is 14) it is then converted into span for loess by dividing by number of rows in df.}

\item{skip_variables}{Variables to skip (default is NULL)}

\item{degree}{Degree of smoothing (default is 1)}

\item{START_DATE}{The date column}
}
\value{
A dataframe with the smoothed data
}
\description{
Smooth double features of a dataframe
}
