% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothGK.R
\name{SmoothGK}
\alias{SmoothGK}
\title{SmoothGK}
\usage{
SmoothGK(x, b = 1)
}
\arguments{
\item{x}{The x vector to smooth. First value is the farthest, last value is the current value.}

\item{b}{The bandwidth (default is 1)}
}
\value{
The smoothed value for the current point.
}
\description{
Smoothing function using gaussian kernel
}
\examples{

x <- 1:5
SmoothGK(x = x, b = 1)

}
