% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabcompareModelCompromise.R
\name{TabcompareModelCompromise}
\alias{TabcompareModelCompromise}
\title{TabcompareModelCompromise}
\usage{
TabcompareModelCompromise(
  bestModelCompromise,
  tabBestModelsPerf,
  dfBestCompromiseModelHierarchy
)
}
\arguments{
\item{bestModelCompromise}{The table of best model compromise}

\item{tabBestModelsPerf}{The table of best model by outcome}

\item{dfBestCompromiseModelHierarchy}{The table of compromise value for each model}
}
\value{
A table with best model and their sum of MAE or MRE over the different outcome times
}
\description{
Generate table of best models with also the bestmodelcompromise
}
