% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabFeatureAutoSelection.R
\name{TabFeatureAutoSelection}
\alias{TabFeatureAutoSelection}
\title{TabFeatureAutoSelection}
\usage{
TabFeatureAutoSelection(dflsDataGroupImp, dflsDataImp)
}
\arguments{
\item{dflsDataGroupImp}{Feature importance from ls_resultsCurtadf_groupVarimp.rds file.}

\item{dflsDataImp}{Feature importance from ls_resultsCurtadf_varimp.rds file.}
}
\value{
A table with the selected features.
}
\description{
Generate table of automated feature selection.
}
