% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabFeatureManualSelection.R
\name{TabFeatureManualSelection}
\alias{TabFeatureManualSelection}
\title{TabFeatureManualSelection}
\usage{
TabFeatureManualSelection(dfDashboard, lsFeatures)
}
\arguments{
\item{dfDashboard}{The dataframe with all features}

\item{lsFeatures}{The list of features}
}
\value{
A table with the feature selection
}
\description{
Get the manual feature selection table
}
