% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabRawVariables.R
\name{TabRawVariables}
\alias{TabRawVariables}
\title{TabRawVariables}
\usage{
TabRawVariables(dfDashboard)
}
\arguments{
\item{dfDashboard}{The dataframe of interest.}
}
\value{
A table of features
}
\description{
Generate table of variables available in dfDashboard
}
