% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainElasticNet.R
\name{trainElasticNet}
\alias{trainElasticNet}
\title{trainElasticNet}
\usage{
trainElasticNet(FeaturesLasso, xTrain, xTest, yTrain)
}
\arguments{
\item{FeaturesLasso}{The Lasso selected features}

\item{xTrain}{The train set}

\item{xTest}{The test set}

\item{yTrain}{The y of train set}
}
\value{
A list of model and predictions
}
\description{
Train elastic-net model with dummies for CHU_WEEKDAY
}
