% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainModelforPublication.R
\name{TrainModelforPublication}
\alias{TrainModelforPublication}
\title{TrainModelforPublication}
\usage{
TrainModelforPublication(lsTrainTest, model, boot = FALSE)
}
\arguments{
\item{lsTrainTest}{A list of length two with test and train sets}

\item{model}{a statistical model either rf, enet (linear elastic-net) or poisson (poisson elastic-net)}

\item{boot}{should bootstrap sampling be performed on train set}
}
\value{
A list of two dataframe, prediction and performance
}
\description{
Train model for publication and return importance and predictions
}
