% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainRf.R
\name{trainRf}
\alias{trainRf}
\title{trainRf}
\usage{
trainRf(FeaturesRf, xTrain, xTest, yTrain)
}
\arguments{
\item{FeaturesRf}{The RF selected features}

\item{xTrain}{The train set}

\item{xTest}{The test set}

\item{yTrain}{The y of train set}
}
\value{
A list of model and predictions
}
\description{
Train randomForest model
}
