% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update2021Weather.R
\name{Update2021Weather}
\alias{Update2021Weather}
\title{Update2021Weather}
\usage{
Update2021Weather(n.cores = 2)
}
\arguments{
\item{n.cores}{Number of cores}
}
\value{
A dataframe with 2020 and 2021 data
}
\description{
Update 2021 weather data and add it to 2020 weather data.
}
