% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDatasetsForPrediction.R
\name{UpdateDatasetsForPrediction}
\alias{UpdateDatasetsForPrediction}
\title{UpdateDatasetsForPrediction}
\usage{
UpdateDatasetsForPrediction(
  dfVariantsPast = readRDS(file = "extdata/dfVariantsImputed.rds"),
  skip_variants = TRUE,
  saveData = FALSE,
  RqCOVID_INDIC,
  RqCOVID_FEATURES_SAVE_BIS
)
}
\arguments{
\item{dfVariantsPast}{Past dataframe of variants with gompertz imputation.}

\item{skip_variants}{Should the variants be skipped (default is TRUE because variants data collection changed).}

\item{saveData}{Should the data be saved.}

\item{RqCOVID_INDIC}{Data from COVID_INDIC table}

\item{RqCOVID_FEATURES_SAVE_BIS}{Data from COVID_FEATURES_SAVE_BIS table}
}
\value{
A list of dataframe
}
\description{
Update data for CHU predictions
}
