% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data_from_NOAA.R
\name{weather_data_from_NOAA}
\alias{weather_data_from_NOAA}
\title{weather_data_from_NOAA}
\usage{
weather_data_from_NOAA(
  Regions_or_Dept_stations_pop,
  years = c(2020, 2021),
  n.cores = 1
)
}
\arguments{
\item{Regions_or_Dept_stations_pop}{The region or departement QGIS dataset available in SEIRcovid19 package.}

\item{years}{Years of interest. Default is 'c(2020, 2021)'.}

\item{n.cores}{Number of cores to download NOAA meteorological dataset.}
}
\value{
A dataframe with 9 columns:
\itemize{
  \item label_insee - The name of the region or the departement.
  \item code_insee - The insee code of the departement or the region.
  \item date_day - The date.
  \item t.mean - The mean temperature over the day..
  \item precip - The precipitation over the day.
  \item RH.mean - The mean relative humidity over the day.
  \item AH.mean - The mean absolute humidity over the day.
  \item IPTCC.mean - The mean IPTCC index over the day.
  \item ws.mean - The mean wind speed over the day.
  \item dewpoint.mean - The mean dew point over the day.
}
}
\description{
Provide the weather data from NOAA. User should provide either 'Regions_stations_pop' or 'Dept_stations_pop' to the Regions_or_Dept_stations_pop parameter which are datasets extracted from QGIS available in SEIRcovid19 package.
}
\examples{
\dontrun{

weather_data_reg <- weather_data_from_NOAA(
 Regions_or_Dept_stations_pop = SEIRcovid19FR::Regions_stations_pop,
 years = 2020:2021,
 n.cores = 2
)

save(weather_data_reg, file = "data/weather_data_reg.rdata")

weather_data_dep <- weather_data_from_NOAA(
 Regions_or_Dept_stations_pop = SEIRcovid19FR::Dept_stations_pop,
 years = 2020:2021,
 n.cores = 2
)

save(weather_data_dep, file = "data/weather_data_dep.rdata")
}
}
